Bump SDK version to 1.5.0
Proposal: keep minor version of SDK in line with minor version of the player. Let’s not do patch versions also, so we can still do fixes without having to bump the player version.

This means that significant fixes in the player (minor version bump) will also be reflected as a significant fix in the SDK.